requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit
openai
pinecone
sentence-transformers
pyyaml
python-dotenv
setuptools==80.3.1
torch
